✨ :  본인 댓글만 삭제 가능

+++ src/main/java/com/springminiprj1/controller/comment/CommentController.java
@@ -3,6 +3,7 @@ package com.springminiprj1.controller.comment;
 import com.springminiprj1.domain.comment.Comment;
 import com.springminiprj1.service.comment.CommentService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
@@ -33,7 +34,12 @@ public class CommentController {
     }
 
     @DeleteMapping("remove")
-    public void remove(@RequestBody Comment comment) {
-        service.deleteById(comment.getId());
+    public ResponseEntity remove(@RequestBody Comment comment, Authentication authentication) {
+        if (service.hasAccess(comment, authentication)) {
+            service.deleteById(comment.getId());
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
     }
 }
